+++ cloudinit/sources/DataSourceSmartOS.py
@@ -30,6 +30,7 @@ import random
 import re
 import socket
 
+from cloudinit import dmi
 from cloudinit import log as logging
 from cloudinit import serial
 from cloudinit import sources
@@ -767,7 +768,7 @@ def get_smartos_environ(uname_version=None, product_name=None):
         return SMARTOS_ENV_LX_BRAND
 
     if product_name is None:
-        system_type = util.read_dmi_data("system-product-name")
+        system_type = dmi.read_dmi_data("system-product-name")
     else:
         system_type = product_name
